avoid unit.wast in wasm-as, as it uses f32.demote/f64 which is not in the binary format currently

+++ check.py
@@ -281,7 +281,7 @@ if torture:
 print '\n[ checking wasm-as testcases... ]\n'
 
 for wast in tests:
-  if wast.endswith('.wast'):
+  if wast.endswith('.wast') and not wast in ['unit.wast']: # blacklist some known failures
     cmd = [os.path.join('bin', 'wasm-as'), os.path.join('test', wast), '-o', 'a.wasm']
     print cmd
     if os.path.exists('a.wasm'): os.unlink('a.wasm')
